apply pr review
